+++ src/Handler/MockHandler.php
@@ -96,8 +96,8 @@ class MockHandler implements \Countable
                 throw new \InvalidArgumentException('on_headers must be callable');
             }
             try {
-                $options['on_headers']($response);
-            } catch (\Exception $e) {
+                $options['on_headers']($response, $request);
+            } catch (\Throwable $e) {
                 $msg = 'An error was encountered during the on_headers event';
                 $response = new RequestException($msg, $request, $response, $e);
             }